+++ Public/PGPFGForPaghiperPixPage.php
@@ -30,11 +30,11 @@ try {
         $transactionId = base64_encode($pixParamsDecoded->transactionId);
         $donationID = base64_encode($pixParamsDecoded->donationId);
 
-        $qrCodeDesc = esc_html__('Chave PIX');
-        $notice = esc_html__('Pix copia e cola, clique no código abaixo para copiar:');
-        $copyMsg = esc_html__('Copiado!');
-        $dueDateMsg = esc_html__(' Vencimento: ');
-        $currencyTxt = esc_html__('R$ 0,00');
+        $qrCodeDesc = esc_html__('Chave PIX', 'payment-gateway-pix-for-givewp');
+        $notice = esc_html__('Pix copia e cola, clique no código abaixo para copiar:', 'payment-gateway-pix-for-givewp');
+        $copyMsg = esc_html__('Copiado!', 'payment-gateway-pix-for-givewp');
+        $dueDateMsg = esc_html__(' Vencimento: ', 'payment-gateway-pix-for-givewp');
+        $currencyTxt = esc_html__('R$ 0,00', 'payment-gateway-pix-for-givewp');
 
         $dueDateMsg .= $donDueDate;